const handleScroll = (id: string) => {     const element = document.getElementById(id)     if (element) {       element.scrollIntoView({         behavior: smooth,         block: start,       })     }   }  const handleScroll = (id: string) => {     const element = document.getElementById(id)     if (element) {       element.scrollIntoView({         behavior: smooth,         block: sta
